добавляю возможность просмотра отчётов по отдельному сотруднику

+++ SalaryProject/Users/Employee.cs
@@ -28,12 +28,12 @@ namespace SalaryProject
             Console.WriteLine("(3). Выход из программы");
         }
 
-        public override int GetSalary(int workTime)
+        public override double GetSalary(double workTime)
         {
-            int salary;
+            double salary;
             if (workTime > 160)
             {
-                int overTime = workTime - 160;
+                double overTime = workTime - 160;
                 salary = 120_000 + (overTime / 160 * 40_000);
             }
             else
